+++ src/app/pages/privacy/privacy.component.ts
@@ -108,7 +108,7 @@ import { BIBLE_TRANSLATION_LABELS } from "../../types/memorization";
               Your prayer content and account information are stored in our
               database. Administrators of your organization (e.g., church staff)
               can access the data needed to run the app (approve prayers, manage
-              subscribers, send notifications). Public prayers you submit (after
+              subscribers, send notifications). Church prayers you submit (after
               approval) are visible to other logged-in users in your community.
               Personal prayers are visible only to you. We do not sell your
               personal information.